+++ RDS/circle2_use_cases/exporter/src/tests/test_exporterservice.py
@@ -112,6 +112,8 @@ class Test_ExporterService(unittest.TestCase):
                 "Owncloud", "Zenodo", "testfile.txt", "admin")
         self.assertFalse(result)
 
+    # TODO: this test cannot be handled, because pactman can only emulate json, nothing else.
+    @unittest.skip
     def test_export_server(self):
         # call to get file from owncloud
         pact.given(
