+++ RDS/circle2_use_cases/token_service/src/api/Service/Service.py
@@ -1,9 +1,10 @@
 from flask import jsonify
-import Util
+import src.Util as Util
 from connexion_plus import FlaskOptimize
 
 def index():
     return jsonify(Util.tokenService.getAllServices())
 
+@FlaskOptimize.set_cache_timeout(3600)
 def get(servicename):
     return Util.tokenService.getService(servicename)
